feat: move city components to pages
